Move gpu/ test folder into correct new place
